Split the instrument snapshot loader from the reference registry
Pure move: InstrumentSnapshotLoader owns the repository reads, map
building, and collision findings, while InstrumentReferenceService
keeps the volatile lookup state, refresh scheduling, and query surface;
the parsing tests move with the loader and the lookup tests run against
the real one.

God classes 10 -> 9, PMD violations 144 -> 142.
