+++ metrics/scorecard.json
@@ -8,14 +8,14 @@
   "disabledTests": 5,
   "disconnectedClasses": 454,
   "fqnViolations": 0,
-  "godClasses": 10,
+  "godClasses": 9,
   "interactionVerifyingTestClasses": 361,
   "interactionVerifyingTestShare": 36.2,
   "lineCoverage": 84.03,
   "longClasses": 3,
   "meanInstability": 0.473,
-  "mocklessTestClasses": 461,
-  "mocklessTestShare": 46.2,
+  "mocklessTestClasses": 462,
+  "mocklessTestShare": 46.3,
   "moduleCycles": 0,
   "moduleSliceTests": 0,
   "modulithViolations": 0,
@@ -26,22 +26,22 @@
   "overMockedTestClasses": 111,
   "overMockedTestShare": 11.1,
   "pitestSlicedModules": 52,
-  "pmdViolations": 144,
+  "pmdViolations": 142,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
     "CouplingBetweenObjects": 35,
-    "CyclomaticComplexity": 75,
-    "GodClass": 10,
+    "CyclomaticComplexity": 74,
+    "GodClass": 9,
     "TooManyMethods": 4
   },
-  "slowest10TestClassesSeconds": 121.3,
+  "slowest10TestClassesSeconds": 118.9,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": null,
+  "suiteWallTimeSeconds": 200,
   "testClassesOver10Seconds": 5,
-  "testClassesTotal": 997,
+  "testClassesTotal": 998,
   "testIntegrationShare": 6.7,
   "testSliceShare": 13.8,
-  "testUnitShare": 79.4,
+  "testUnitShare": 79.5,
   "top15ClassLines": 6875,
   "totalPackages": 227,
   "unmarkedPackages": 0
